# public/swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Online hautarzt API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/enquires/{id}/verify-sms": {
            "post": {
                "tags": [
                    "Enquires"
                ],
                "summary": "Verify SMS verification code",
                "description": "Verify SMS verification code",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "An enquire's identificator",
                        "required": true,
                        "example": "1"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "verification_code",
                                    "recaptcha"
                                ],
                                "properties": {
                                    "verification_code": {
                                        "title": "Verification code",
                                        "description": "Verification code",
                                        "format": "string",
                                        "example": "95843565"
                                    },
                                    "recaptcha": {
                                        "title": "Recaptcha",
                                        "description": "A recaptcha token. The action must be 'verify_sms'",
                                        "format": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Verification code has been verified",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "format": "object",
                                            "$ref": "#/components/schemas/AuthTokenResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "There are some validation errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "title": "Validation error",
                                    "properties": {
                                        "message": {
                                            "format": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "properties": {
                                                "validation_code": {
                                                    "items": {
                                                        "type": "string",
                                                        "example": "Verification code is invalid."
                                                    }
                                                }
                                            },
                                            "type": "object",
                                            "format": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Resource not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "format": "string",
                                            "example": "No query results for model [App\\Models\\Enquire]."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal technical error was happened",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "format": "string",
                                            "example": "Something went wrong, please try again later."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "AuthTokenResource": {
                "title": "AuthToken resource",
                "description": "Resource for an authorization token representation",
                "properties": {
                    "id": {
                        "format": "integer",
                        "example": "1"
                    },
                    "access_token": {
                        "format": "string",
                        "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJSUzI1NiIsImp0aSI6IjQ3ZGY5ZDdkYmY4ZmM1Mz"
                    },
                    "expires_at": {
                        "format": "object",
                        "$ref": "#/components/schemas/CarbonResource"
                    }
                },
                "type": "object"
            },
...
            "CarbonResource": {
                "title": "Carbon resource",
                "description": "Resource for datetime representation",
                "properties": {
                    "date": {
                        "format": "string",
                        "example": "2019-12-11 12:25:00.000000"
                    },
                    "timezone_type": {
                        "format": "int64",
                        "example": 3
                    },
                    "timezone": {
                        "format": "string",
                        "example": "Europe/Berlin"
                    }
                },
                "type": "object"
            },
...
        }
    }
}